emote scraper rewrite
The emote scraper is now multithreaded ans _should_ be faster.
Also, the class based architecture allows easy extensibility for future
projects. To process the emotes for a different project another superclass of
AbstractEmoteProcessor can be added.
The scraper also does proper rate limiting, if it is configured to do so.